refactor(convert): extract list_marker module
Phase 1 of fulgur-odgo. Move resolve_list_style_image_asset,
size_raster_marker, resolve_list_marker, resolve_inside_image_marker,
extract_marker_lines, find_marker_font, shape_marker_with_skrifa,
inject_inside_marker_item_into_children from convert/mod.rs into
convert/list_marker.rs. has_paragraph_descendant in mod.rs becomes
pub(super) so the moved fn can call back into it.

No behavioral change. VRT byte-identical against committed goldens.
